pyproject.toml
[project]
name = "github-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.2.0",
    "langchain>=1.2.2",
    "langchain-anthropic>=1.3.1",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.0",
    "pinecone>=7.3.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
]
